Feature FE delete order by user 구현

See merge request sw_track/class_03/web_project/team4/eladin!66

+++ src/db/models/order-model.js
@@ -123,6 +123,10 @@ export class OrderModel {
     }
     
     async cancelOrderById(orderId) {
+        const targetOrder = await Order.findById({_id : orderId}).exec();
+        if(targetOrder.shippingStatus !== '배송준비중'){
+            throw new Error(`배송준비중인 주문만 취소가 가능합니다.`)    
+        }
         await Order.deleteOne({ _id : orderId })
     }
